# internal/docs/swagger.yaml
definitions:
...
  handler.createDoctorRequest:
    properties:
      bio:
        type: string
      designation:
        type: string
      education:
        type: string
      experience:
        type: integer
      license_no:
        type: string
      specialty:
        type: string
    required:
    - license_no
    - specialty
    type: object
...
  handler.doctorResponse:
    properties:
      bio:
        type: string
      designation:
        type: string
      education:
        type: string
      email:
        type: string
      experience:
        type: integer
      id:
        type: integer
      license_no:
        type: string
      name:
        type: string
      specialty:
        type: string
      user_id:
        type: integer
    type: object
...
  handler.updateDoctorRequest:
    properties:
      bio:
        type: string
      designation:
        type: string
      education:
        type: string
      experience:
        type: integer
      license_no:
        type: string
      specialty:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /doctors:
    get:
      description: Get a paginated list of all doctors
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of doctors
          schema:
            items:
              $ref: '#/definitions/handler.doctorResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all doctors
      tags:
      - doctors
    post:
      consumes:
      - application/json
      description: Create a new doctor profile
      parameters:
      - description: Doctor Information
        in: body
        name: doctor
        required: true
        schema:
          $ref: '#/definitions/handler.createDoctorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created doctor profile
          schema:
            $ref: '#/definitions/handler.doctorResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create doctor profile
      tags:
      - doctors
...
  /doctors/{id}:
    delete:
      description: Delete a doctor profile by ID
      parameters:
      - description: Doctor ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete doctor profile
      tags:
      - doctors
    get:
      description: Get a doctor profile by ID
      parameters:
      - description: Doctor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Doctor profile
          schema:
            $ref: '#/definitions/handler.doctorResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get doctor profile
      tags:
      - doctors
    put:
      consumes:
      - application/json
      description: Update an existing doctor profile
      parameters:
      - description: Doctor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Doctor Information
        in: body
        name: doctor
        required: true
        schema:
          $ref: '#/definitions/handler.updateDoctorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated doctor profile
          schema:
            $ref: '#/definitions/handler.doctorResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update doctor profile
      tags:
      - doctors
  /doctors/specialty/{specialty}:
    get:
      description: Get a paginated list of doctors by specialty
      parameters:
      - description: Specialty
        in: path
        name: specialty
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of doctors
          schema:
            items:
              $ref: '#/definitions/handler.doctorResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List doctors by specialty
      tags:
      - doctors
  /doctors/user/{userId}:
    get:
      description: Get a doctor profile by user ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Doctor profile
          schema:
            $ref: '#/definitions/handler.doctorResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get doctor profile by user ID
      tags:
      - doctors
...
swagger: "2.0"